Allow configuration of StatsD metrics backend
If either `SENTRY_STATSD_HOST` or `SENTRY_STATSD_PORT` is defined as environment
variables the StatsD metrics backend will be enabled and use the provided
settings to the configuration. If either the host or port is not defined they
will default to use a hostname of 'localhost' and port of 8125.
https://docs.sentry.io/server/internal-metrics/