Allow configuration of DogStatsD metrics backend
If either `SENTRY_DOGSTATSD_HOST` or `SENTRY_DOGSTATSD_PORT` is defined as
environment variables the DogStatsD metrics backend will be enabled and use the
provided settings to the configuration. If either the host or port is not
defined they will default to use a hostname of 'localhost' and port of 8125.
https://docs.sentry.io/server/internal-metrics/

The parameters mentioned in the above documentation for setting the hostname
and port are wrong, https://github.com/getsentry/sentry-docs/pull/465 corrects
that.